ansible/requirements.txt
ansible==2.8.1
ansible-lint==4.1.0
anyconfig==0.9.7
aodhclient==1.2.0
appdirs==1.4.3
arrow==0.14.2
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
Babel==2.7.0
binaryornot==0.4.4
Cerberus==1.2
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
click==6.7
click-completion==0.3.1
cliff==2.14.1
cmd2==0.9.12
colorama==0.3.9
cookiecutter==1.6.0
cryptography==2.7
debtcollector==1.21.0
decorator==4.4.0
dogpile.cache==0.7.1
entrypoints==0.3
fasteners==0.15
flake8==3.7.7
future==0.17.1
futurist==1.8.1
git-url-parse==1.2.2
gnocchiclient==7.0.5
idna==2.7
importlib-metadata==0.18
iso8601==0.1.12
Jinja2==2.10.1
jinja2-time==0.2.0
jmespath==0.9.4
jsonpatch==1.23
jsonpointer==2.0
jsonschema==3.0.1
keystoneauth1==3.14.0
MarkupSafe==1.1.1
mccabe==0.6.1
minimal==0.1.0
molecule==2.20.1
monotonic==1.5
more-itertools==7.0.0
msgpack==0.6.1
munch==2.3.2
murano-pkg-check==0.3.0
netaddr==0.7.19
netifaces==0.10.9
networkx==2.3
openstackclient==0.1.0
openstacksdk==0.31.1
os-client-config==1.32.0
os-service-types==1.7.0
osc-lib==1.12.1
oslo.config==6.9.0
oslo.context==2.22.1
oslo.i18n==3.23.1
oslo.log==3.44.0
oslo.serialization==2.29.1
oslo.utils==3.41.0
packaging==19.0
passlib==1.7.1
pathspec==0.5.9
pbr==5.1.1
pexpect==4.6.0
pluggy==0.12.0
ply==3.11
poyo==0.4.2
prettytable==0.7.2
psutil==5.4.6
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pydot==1.4.1
pyflakes==2.1.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyperclip==1.7.0
pyrsistent==0.15.2
pytest==4.6.3
python-barbicanclient==4.8.1
python-cinderclient==4.2.0
python-congressclient==1.12.0
python-dateutil==2.8.0
python-designateclient==2.11.0
python-gilt==1.2.1
python-glanceclient==2.16.0
python-heatclient==1.17.0
python-ironic-inspector-client==3.5.0
python-ironicclient==2.8.0
python-keystoneclient==3.19.0
python-mistralclient==3.9.0
python-muranoclient==1.2.0
python-neutronclient==6.12.0
python-novaclient==14.1.0
python-octaviaclient==1.9.0
python-openstackclient==3.19.0
python-saharaclient==2.2.0
python-searchlightclient==1.5.1
python-senlinclient==1.10.1
python-swiftclient==3.7.0
python-troveclient==2.18.0
python-vitrageclient==2.7.0
python-watcherclient==2.2.0
python-zaqarclient==1.11.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requestsexceptions==1.4.0
rfc3986==1.3.2
ruamel.yaml==0.15.97
semantic-version==2.6.0
sh==1.12.14
simplejson==3.16.0
six==1.11.0
stevedore==1.30.1
tabulate==0.8.2
testinfra==3.0.5
tree-format==0.1.2
ujson==1.35
urllib3==1.25.3
warlock==1.3.3
wcwidth==0.1.7
whichcraft==0.5.2
wrapt==1.11.1
yamllint==1.16.0
yaql==1.1.3
zipp==0.5.1
